requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
pydantic>=2.8.0
pandas>=2.2.0
pyarrow>=16.0.0
datasets>=2.20.0
python-dotenv>=1.0.1
rapidfuzz>=3.9.0
pytest>=8.2.0
groq>=0.11.0
